storage-api/pyproject.toml
[tool.poetry]
name = "storage-api"
version = "0.1.0"
description = ""
authors = []
license = "Proprietary"


[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
